finally found reason for why pillars doesn't take damage!
the switch to quadtree space instead of simple or hash cause ode to
most of the time not generate any forces to geoms with no bodies.
this is a good thing, since geoms with no body doesn't respond to forces
anyway (should be the default, and don't know why switch of space change
it). but I originally sent the force on geom1 to geom1, and the same
for geom2. Now the biggest force is calculated, and sent to both geoms.